% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja wykonująca symulację zadanej strategii określoną liczbę razy.}
\usage{
badaj_gre(strategia, liczba_iteracji = 1000)
}
\arguments{
\item{strategia}{- argument określający funkcję strategii gry}

\item{liczba_iteracji}{- argument okreslajacy dla ilu wywolan gry chcemy otrzymac
wynik funkcji. Domyslna wartosc to 10000.}
}
\value{
funkcja zwraca wektor długości \code{liczba_iteracji} czasów kolejnych rozgrywek.
}
\description{
Funcja \code{badaj_gre} symuluje \code{liczba_iteracji} rozgrywek
i zwraca wektor czasów rozgrywki.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

