% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czy_koniec.R
\name{czy_koniec}
\alias{czy_koniec}
\title{Funkcja prywatna sprawdzająca warunek zakończenia gry.}
\usage{
czy_koniec(zagroda)
}
\arguments{
\item{zagroda}{- argument określający obecny stan zagrody gracza.}
}
\description{
Funkcja \code{czy_koniec} zwraca \code{true} wtedy i tylko wtedy, gdy spełniony jest warunek
zakończenia rozgrywki i \code{false} w przeciwnym razie.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

