% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Funkcja wykonująca pojedynczą symulację zadanej strategii}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{- argument określający funkcję strategii gry}
}
\value{
funkcja zwraca czas trwania rozgrywki podny w liczbie rund potrzebnych do uzyskania
przez gracza zagrody spełniającej warunek końcowy.
}
\description{
Funcja \code{gra} symuluje pojedynczą rozgrywkę i zwraca czas jej trwania.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

