% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indeks.R
\name{indeks}
\alias{indeks}
\title{Funkcja zwracająca indeks zwierzęcia}
\usage{
indeks(zwierze)
}
\arguments{
\item{zwierze}{- parametr okreslający zwierze}
}
\value{
funkcja zwraca uniwersalny indeks danego zwierzęcia
}
\description{
Funkcja zwracająca indeks zwierzęcia
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

