% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kup_kilka_tanszych.R
\name{kup_kilka_tanszych}
\alias{kup_kilka_tanszych}
\title{Funkcja wymieniająca droższe zwierzę na kilka tańszych}
\usage{
kup_kilka_tanszych(zagroda, zwierze)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbê poszczególnych rodzajów zwierząt w zagrodzie}

\item{zwierze}{- numer zwierzęcia, które chcemy sprzedać}
}
\value{
wektor reprezentujący stan zagrody po dokonianiu odpowiednich wymian
}
\description{
Jeśli nie mamy żadnego egzemplarza któregoś ze zwierząt tañszych niż zwierzę, które chcemy wymienić, to po kolei je kupujemy za sprzedane zwierzę.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

