% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prawie_wygrana.R
\name{prawie_wygrana}
\alias{prawie_wygrana}
\title{Funkcja wykonująca część strategii dla momentu, kiedy wartość zagrody jest bliska 127}
\usage{
prawie_wygrana(zagroda)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbę poszczególnych rodzajów zwierząt w zagrodzie}
}
\value{
wektor reprezentujący stan zagrody po dokonianiu odpowiednich wymian
}
\description{
Dokonuje odpowiednich wymian niezbędnych do uzyskania co najmniej jednego egzemplarza zwierzęcia z każdego gatunku.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

