% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_histogram.R
\name{rysuj_histogram}
\alias{rysuj_histogram}
\title{Funkcja generuje histogram z wyniku funkcji badaj_grę}
\usage{
rysuj_histogram(wyniki)
}
\arguments{
\item{wyniki}{- argument określający wyniki badania strategii}
}
\description{
Funkcja generuje histogram z wyniku funkcji badaj_grę
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

