% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzuc_kostkami.R
\name{rzuc_kostkami}
\alias{rzuc_kostkami}
\title{Funkcja wykonująca rzut dwiema kostkami określonymi w zasadach gry.}
\usage{
rzuc_kostkami()
}
\value{
funkcja zwraca wektor długości dwa z wartościomi określającymi wyniki na obydwu kostkach
}
\description{
Bezparametrowa funkcja \code{rzuc_kostkami} zwraca wektor wyników rzutu na dwóch kostkach.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

