% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprzedaj_drugiego_konia.R
\name{sprzedaj_drugiego_konia}
\alias{sprzedaj_drugiego_konia}
\title{Funkcja wymieniająca drugiego konia na inne zwierzęta}
\usage{
sprzedaj_drugiego_konia(zagroda)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbę poszczególnych rodzajów zwierząt w zagrodzie}
}
\value{
wektor reprezentujący stan zagrody po dokonianiu wymiany
}
\description{
Funcja wymienia drugiego konia po kolei na niezbędne do wygranej mniej wartościowe zwierzęta
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

