% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprzedaj_kilka_tanszych.R
\name{sprzedaj_kilka_tanszych}
\alias{sprzedaj_kilka_tanszych}
\title{Funkcja wymieniająca kilka tańszych zwierząt na jedno droższe.}
\usage{
sprzedaj_kilka_tanszych(zagroda, zwierze)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbę poszczególnych rodzajów zwierząt w zagrodzie}

\item{zwierze}{- numer zwierzęcia, które chcemy kupić}
}
\value{
wektor reprezentujący stan zagrody po dokonianiu odpowiednich wymian
}
\description{
Jeśli jesteœmy w stanie sobie pozwoliæ na kupowane zwierzê i zachować po tej transakcji co najmniej jeden egzemplarz ze wszytkich wcześniej posiadanych zwierząt, to funkcja kupuje to zwierzę.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

