% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_asdf.R
\name{strategia_asdf}
\alias{strategia_asdf}
\title{Podstawowa strategia rozgrywnia partii (niezgodna z zasadami)}
\usage{
strategia_asdf(zagroda)
}
\arguments{
\item{zagroda}{- argument będący wektorym określającym obecny stan zagrody gracza}
}
\value{
funkcja zwraca nowy wektor \code{zagroda} po wykonaniu przez gracza ruchu.
}
\description{
Funcja \code{strategia_asdf} dla dowolnego stanu zagrody zwraca zagrodę w stanie wygrywającym.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

