% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_p.R
\name{strategia_p}
\alias{strategia_p}
\title{Podstawowa strategia rozgrywnia partii}
\usage{
strategia_p(zagroda)
}
\arguments{
\item{zagroda}{- argument będący wektorym określającym obecny stan zagrody gracza}
}
\value{
funkcja zwraca nowy wektor \code{zagroda} po wykonaniu przez gracza ruchu.
}
\description{
Funcja \code{strategia_p} dla dowolnego stanu zagrody zwraca zagrodę w następnym stanie.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

