% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_asdf.R
\name{strategy_asdf}
\alias{strategy_asdf}
\title{Podstawowa strategia rozgrywnia partii}
\usage{
strategy_asdf(zagroda)
}
\arguments{
\item{zagroda}{- argument będący wektorym określającym obecny stan zagrody gracza}
}
\value{
funkcja zwraca wektor długości \code{liczba.iteracji} czasów kolejnych rozgrywek.
}
\description{
Funcja \code{strategy asdf} dla każdego zadanego stanu zagrody gracza
zwraca stan zagrody po wykonaniu przez gracza ruchu.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}
