% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartosc.R
\name{wartosc}
\alias{wartosc}
\title{Funkcja zwracająca wartość danego zwierzaka}
\usage{
wartosc(zwierze)
}
\arguments{
\item{zwierze}{- argument określający typ zwierzaka.}
}
\value{
Funkcja zwraca wartość podanego zwierzaka.
}
\description{
Funkcja \code{wartosc} zwraca wartość zwierzeka podanego jako argument.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

