% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartosc_zagrody.R
\name{wartosc_zagrody}
\alias{wartosc_zagrody}
\title{Funkcja obliczająca łączną wartość zagrody w królikach.}
\usage{
wartosc_zagrody(zagroda)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbę poszczególnych rodzajów zwierząt w zagrodzie}
}
\value{
wartość wszystkich zwierząt w zagrodzie wyrażona w liczbie królików
}
\description{
Funcja mnoży wektor reprezentujący zwierzeta w zagrodzie przez wektor ich wartoœci w królikach.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

