% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykonaj_kostki.R
\name{wykonaj_kostki}
\alias{wykonaj_kostki}
\title{Funkcja wprowadzająca zmiany w zagrodzie na podstawie wyniku uzyskanego na kostkach.}
\usage{
wykonaj_kostki(zagroda, wynik)
}
\arguments{
\item{zagroda}{- argument określający obecny stan zagrody gracza.}

\item{wynik}{- argument określający wynik losowania na kostkach (wektor długości 2).}
}
\value{
funkcja zwraca nowy stan zagrody gracza
}
\description{
Funkcja zmienia liczebność zagrody gracza na podstawie uzyskanych wyniów na kostkach
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

