% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wymiana.R
\name{wymiana}
\alias{wymiana}
\title{Funkcja wymieniająca zwierzę na inne}
\usage{
wymiana(zagroda, co_sprzedaj, co_kup, ile_sprzedaj, ile_kup)
}
\arguments{
\item{zagroda}{- wektor reprezentujący liczbę poszczególnych rodzajów zwierząt w zagrodzie}

\item{co_sprzedaj}{- nazwa zwierzęcia sprzedawanego}

\item{co_kup}{- numer zwierzęcia kupowanego}

\item{ile_sprzedaj}{- liczba zwierząt sprzedawanych}

\item{ile_kup}{- liczba zwierząt kupowanych}
}
\value{
wektor reprezentujący stan zagrody po dokonianiu wymiany
}
\description{
Funcja dodaje liczbę kupionych do zagrody i odejmuje jej równowartość w sprzedawanych zwierzętach.
}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

